pyproject.toml
[build-system]
requires = ["hatchling>=1.25"]
build-backend = "hatchling.build"

[project]
name = "konnaxion-capsule-manager"
version = "0.1.0"
description = "Konnaxion Capsule Manager, Agent, Builder, and CLI"
readme = "README.md"
requires-python = ">=3.11"
license = { text = "Proprietary" }
authors = [
  { name = "Konnaxion" }
]
keywords = [
  "konnaxion",
  "capsule",
  "manager",
  "agent",
  "docker-compose",
  "appliance"
]
classifiers = [
  "Development Status :: 3 - Alpha",
  "Programming Language :: Python :: 3",
  "Programming Language :: Python :: 3.11",
  "Programming Language :: Python :: 3.12",
  "Private :: Do Not Upload"
]

dependencies = [
  "cryptography>=42.0",
  "docker>=7.0",
  "fastapi>=0.111",
  "httpx>=0.27",
  "pydantic>=2.7",
  "pydantic-settings>=2.3",
  "python-dotenv>=1.0",
  "pyyaml>=6.0",
  "rich>=13.7",
  "typer>=0.12",
  "uvicorn[standard]>=0.30",
  "zstandard>=0.22"
]

[project.optional-dependencies]
dev = [
  "mypy>=1.10",
  "pytest>=8.2",
  "pytest-cov>=5.0",
  "ruff>=0.5",
  "types-PyYAML>=6.0"
]

[project.scripts]
kx = "kx_cli.main:app"
kx-agent = "kx_agent.main:app"
kx-manager = "kx_manager.main:app"
kx-builder = "kx_builder.main:app"

[tool.hatch.build.targets.wheel]
packages = [
  "kx_shared",
  "kx_agent",
  "kx_manager",
  "kx_builder",
  "kx_cli"
]

[tool.hatch.build.targets.sdist]
include = [
  "kx_shared",
  "kx_agent",
  "kx_manager",
  "kx_builder",
  "kx_cli",
  "templates",
  "profiles",
  "policies",
  "tests",
  "README.md",
  "pyproject.toml"
]

[tool.pytest.ini_options]
minversion = "8.0"
addopts = "-ra --strict-markers --strict-config"
testpaths = ["tests"]
python_files = ["test_*.py"]
python_classes = ["Test*"]
python_functions = ["test_*"]

[tool.ruff]
target-version = "py311"
line-length = 100
src = [
  "kx_shared",
  "kx_agent",
  "kx_manager",
  "kx_builder",
  "kx_cli",
  "tests"
]

[tool.ruff.lint]
select = [
  "E",
  "F",
  "I",
  "B",
  "UP",
  "SIM",
  "RUF"
]
ignore = [
  "E501"
]

[tool.ruff.format]
quote-style = "double"
indent-style = "space"
line-ending = "lf"

[tool.mypy]
python_version = "3.11"
strict = true
warn_unused_configs = true
warn_return_any = true
warn_unreachable = true
disallow_untyped_defs = true
disallow_incomplete_defs = true
no_implicit_optional = true
check_untyped_defs = true
packages = [
  "kx_shared",
  "kx_agent",
  "kx_manager",
  "kx_builder",
  "kx_cli"
]

[[tool.mypy.overrides]]
module = [
  "docker.*"
]
ignore_missing_imports = true

[tool.coverage.run]
branch = true
source = [
  "kx_shared",
  "kx_agent",
  "kx_manager",
  "kx_builder",
  "kx_cli"
]

[tool.coverage.report]
show_missing = true
skip_covered = true
fail_under = 80
